Moves entities package into data package
