+++ app/src/main/java/angelosz/catbattlegame/ui/battlechests/BattleChestsUiState.kt
@@ -3,8 +3,8 @@ package angelosz.catbattlegame.ui.battlechests
 import angelosz.catbattlegame.domain.enums.BattleChestType
 import angelosz.catbattlegame.domain.enums.CatRarity
 import angelosz.catbattlegame.domain.enums.ScreenState
-import angelosz.catbattlegame.domain.models.entities.BattleChest
-import angelosz.catbattlegame.domain.models.entities.Cat
+import angelosz.catbattlegame.data.entities.BattleChest
+import angelosz.catbattlegame.data.entities.Cat
 
 data class BattleChestsUiState(
     val battleChests: Map<Pair<CatRarity, BattleChestType>, List<BattleChest>> = mapOf(),
