+++ app/src/main/java/angelosz/catbattlegame/ui/theme/Theme.kt
@@ -94,7 +94,7 @@ fun CatBattleGameTheme(
             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
         }*/
 
-        darkTheme -> DarkColorScheme
+        //darkTheme -> DarkColorScheme
         else -> LightColorScheme
     }
